src/eraven_gig000_venv/requirements.txt
django
django-environ
django-ratelimit
djangorestframework_simplejwt
Pillow
requests
pyjwt
psycopg2